Fix according review
- Rename some classes, now they begin from `Test` as
  described in readmy.
- Add comments to new implemented tests.
- Use h2_connection method `increment_flow_control_window`
- Move some tests from test_h2_block_action to test_h2_frames
